Some refactoring in the playes class

+++ src/constants/CarConstants.java
@@ -4,5 +4,5 @@ public class CarConstants {
 
     public static final String CAR_STRING = "car";
     public static final long SECONDS_DURATION = 60;
-    public static final double PLAYER_CAR_ACCELERATION_OFFSET = 0.1;
+    public static final float PLAYER_CAR_ACCELERATION_OFFSET = 0.1f;
 }
